separo la clase en su propio modulo
